refactor cmd folder: - move main inside another inner notifier folder - move logging from cmd to common
