+++ cmd/notifier/main.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/ElrondNetwork/elrond-go-core/core/check"
 	logger "github.com/ElrondNetwork/elrond-go-logger"
-	"github.com/ElrondNetwork/notifier-go/cmd/logging"
+	"github.com/ElrondNetwork/notifier-go/common/logging"
 	"github.com/ElrondNetwork/notifier-go/config"
 	"github.com/ElrondNetwork/notifier-go/notifier"
 	"github.com/urfave/cli"
